Store invoice amount as decimal (major units); auto-migration on startup

+++ app/models/invoice.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Numeric
 from datetime import datetime
 from app.db.base import Base
 
@@ -21,8 +21,8 @@ class Invoice(Base):
     provider_invoice_id = Column(String, nullable=True, index=True, unique=True)
     provider_payment_id = Column(String, nullable=True, index=True, unique=True)
 
-    # Amounts are stored in minor units (e.g. cents)
-    amount = Column(Integer, nullable=False)  # total amount in minor units
+    # Amount in major units (e.g. 11.81 for SGD 11.81); stored as decimal for precision
+    amount = Column(Numeric(12, 2), nullable=False)
     currency = Column(String, nullable=False)
 
     status = Column(String, nullable=False, default="succeeded")
